requirements.txt
Django==3.2.13
psycopg2-binary==2.9.3
gunicorn==20.1.0
django-versatileimagefield==2.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
django-templated-email==3.0.0
google-i18n-address==2.5.1
django-cors-headers==3.12.0
mysqlclient==2.1.1